combine accessibility tests
